backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.1.1
email-validator==2.1.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
celery==5.3.4
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.15.0
statsmodels==0.14.0
httpx==0.25.2
groq==0.4.1
openai==1.3.5
transformers==4.35.2
torch==2.1.1
nltk==3.8.1
textblob==0.17.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0
aioredis==2.0.1
httpx==0.25.2
apscheduler==3.10.4
pytest==7.4.3
pytest-asyncio==0.21.1
